// types/log.types.ts
export enum LogLevel {
  DEBUG = "DEBUG",
  INFO = "INFO",
  WARN = "WARN",
  ERROR = "ERROR",
}

export enum LogAction {
  UPLOAD = "UPLOAD",
  DOWNLOAD = "DOWNLOAD",
  DELETE = "DELETE",
  AUTH_LOGIN = "AUTH_LOGIN",
  AUTH_LOGOUT = "AUTH_LOGOUT",
  AUTH_FAILED = "AUTH_FAILED",
  FILE_EXPIRED = "FILE_EXPIRED",
  USER_CREATED = "USER_CREATED",
  USER_DELETED = "USER_DELETED",
  ADMIN_ACTION = "ADMIN_ACTION",
  CREATE_SERVICE = "CREATE_SERVICE",
  DELETE_SERVICE = "DELETE_SERVICE",
  UPDATE_SERVICE = "UPDATE_SERVICE",
  SERVICE_UPLOAD = "SERVICE_UPLOAD",
  SERVICE_DOWNLOAD = "SERVICE_DOWNLOAD",
  FOLDER_RENAMED = "FOLDER_RENAMED",
  FOLDER_DELETED = "FOLDER_DELETED",
  SERVICE_INVITE_CREATED = "SERVICE_INVITE_CREATED",
  SERVICE_INVITE_USED = "SERVICE_INVITE_USED",
  ACCESS_TOKEN_CREATED = "ACCESS_TOKEN_CREATED",
  ACCESS_TOKEN_DELETED = "ACCESS_TOKEN_DELETED",
}